+++ src/apps/quotes/apps.py
@@ -1,0 +1,10 @@
+"""Quotes app config."""
+
+from django.apps import AppConfig
+
+
+class QuotesConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.quotes"
+    label = "quotes"
+    verbose_name = "Cotizaciones"